2020-04-10: Prevent negative rating punishment
Players can be moved to sideEnemy[1] for killing civilians (bad),
teamkilling (regrettable, but forgivable), or getting their helicopters
destroyed by the enemy for some reason (not good).  According to the
BI wiki, this happens when player rating[2] drops below -2000[1][3].
This hotfix adds a HandleRating EH[4] which caps the rating at 0.

---
[1] https://community.bistudio.com/wiki/sideEnemy
[2] https://community.bistudio.com/wiki/rating
[3] https://community.bistudio.com/wiki/ArmA:_Rating_Values
[4] https://community.bistudio.com/wiki/Arma_3:_Event_Handlers#HandleRating